bloomTechfetchNews/requirements.txt
azure-functions
requests
beautifulsoup4

